# server/controllers/Skaffolder_shalcom_OBP_db/generated/AccountControllerGenerated.js
/**
 * 
 * 
  _____                      _              _ _ _     _   _     _        __ _ _      
 |  __ \                    | |            | (_) |   | | | |   (_)      / _(_) |     
 | |  | | ___    _ __   ___ | |_    ___  __| |_| |_  | |_| |__  _ ___  | |_ _| | ___ 
 | |  | |/ _ \  | '_ \ / _ \| __|  / _ \/ _` | | __| | __| '_ \| / __| |  _| | |/ _ \
 | |__| | (_) | | | | | (_) | |_  |  __/ (_| | | |_  | |_| | | | \__ \ | | | | |  __/
 |_____/ \___/  |_| |_|\___/ \__|  \___|\__,_|_|\__|  \__|_| |_|_|___/ |_| |_|_|\___|
 
 * DO NOT EDIT THIS FILE!! 
 * 
 *  TO CUSTOMIZE AccountControllerGenerated.js PLEASE EDIT ../AccountController.js
 * 
 *  -- THIS FILE WILL BE OVERWRITTEN ON THE NEXT SKAFFOLDER'S CODE GENERATION --
 * 
 */
import Properties from "../../../properties";
import AccountModel from "../../../models/Skaffolder_shalcom_OBP_db/AccountModel";
import ErrorManager from "../../../classes/ErrorManager";
import { authorize } from "../../../security/SecurityManager";
import AccountController from "../AccountController";

const generatedControllers = {
  /**
   * Init routes
   */
  init: router => {
    const baseUrl = `${Properties.api}/account`;
    router.post(baseUrl + "", authorize([]), AccountController.create);
    router.delete(baseUrl + "/:id", authorize([]), AccountController.delete);
    router.get(baseUrl + "/:id", authorize([]), AccountController.get);
    router.get(baseUrl + "", authorize([]), AccountController.list);
    router.post(baseUrl + "/:id", authorize([]), AccountController.update);
  },


  // CRUD METHODS


  /**
  * AccountModel.create
  *   @description CRUD ACTION create
  *   @param Account obj Object to insert
  *
  */
  create: async (req, res) => {
    try {
      const result = await AccountModel.create(req.body);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * AccountModel.delete
  *   @description CRUD ACTION delete
  *   @param ObjectId id Id Account
  *
  */
  delete: async (req, res) => {
    try {
      const result = await AccountModel.delete(req.params.id);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * AccountModel.get
  *   @description CRUD ACTION get
  *   @param ObjectId id Id Account
  *   @returns Account
  *
  */
  get: async (req, res) => {
    try {
      const result = await AccountModel.get(req.params.id);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * AccountModel.list
  *   @description CRUD ACTION list
  *   @returns ARRAY OF Account
  *
  */
  list: async (req, res) => {
    try {
      const result = await AccountModel.list();
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  
  /**
  * AccountModel.update
  *   @description CRUD ACTION update
  *   @param ObjectId id Id Account
  *   @returns Account
  *
  */
  update: async (req, res) => {
    try {
      const result = await AccountModel.update(req.body);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  
  // Custom APIs

};

export default generatedControllers;