# server/controllers/Skaffolder_shalcom_OBP_db/generated/Customer_accountControllerGenerated.js
/**
 * 
 * 
  _____                      _              _ _ _     _   _     _        __ _ _      
 |  __ \                    | |            | (_) |   | | | |   (_)      / _(_) |     
 | |  | | ___    _ __   ___ | |_    ___  __| |_| |_  | |_| |__  _ ___  | |_ _| | ___ 
 | |  | |/ _ \  | '_ \ / _ \| __|  / _ \/ _` | | __| | __| '_ \| / __| |  _| | |/ _ \
 | |__| | (_) | | | | | (_) | |_  |  __/ (_| | | |_  | |_| | | | \__ \ | | | | |  __/
 |_____/ \___/  |_| |_|\___/ \__|  \___|\__,_|_|\__|  \__|_| |_|_|___/ |_| |_|_|\___|
 
 * DO NOT EDIT THIS FILE!! 
 * 
 *  TO CUSTOMIZE customer_accountControllerGenerated.js PLEASE EDIT ../customer_accountController.js
 * 
 *  -- THIS FILE WILL BE OVERWRITTEN ON THE NEXT SKAFFOLDER'S CODE GENERATION --
 * 
 */
import Properties from "../../../properties";
import Customer_accountModel from "../../../models/Skaffolder_shalcom_OBP_db/Customer_accountModel";
import ErrorManager from "../../../classes/ErrorManager";
import { authorize } from "../../../security/SecurityManager";
import Customer_accountController from "../Customer_accountController";

const generatedControllers = {
  /**
   * Init routes
   */
  init: router => {
    const baseUrl = `${Properties.api}/customer_account`;
    router.post(baseUrl + "", authorize([]), Customer_accountController.create);
    router.delete(baseUrl + "/:id", authorize([]), Customer_accountController.delete);
    router.get(baseUrl + "/findByaccount_id/:key", authorize([]), Customer_accountController.findByaccount_id);
    router.get(baseUrl + "/:id", authorize([]), Customer_accountController.get);
    router.get(baseUrl + "", authorize([]), Customer_accountController.list);
    router.post(baseUrl + "/:id", authorize([]), Customer_accountController.update);
  },


  // CRUD METHODS


  /**
  * customer_accountModel.create
  *   @description CRUD ACTION create
  *   @param customer_account obj Object to insert
  *
  */
  create: async (req, res) => {
    try {
      const result = await Customer_accountModel.create(req.body);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * customer_accountModel.delete
  *   @description CRUD ACTION delete
  *   @param ObjectId id Id customer_account
  *
  */
  delete: async (req, res) => {
    try {
      const result = await Customer_accountModel.delete(req.params.id);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * customer_accountModel.findByaccount_id
  *   @description CRUD ACTION findByaccount_id
  *   @param Objectid key Id of the resource account_id to search
  *
  */
  findByaccount_id: async (req, res) => {
    try {
      const result = await Customer_accountModel.findByaccount_id(req.params.key);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * customer_accountModel.get
  *   @description CRUD ACTION get
  *   @param ObjectId id Id customer_account
  *   @returns customer_account
  *
  */
  get: async (req, res) => {
    try {
      const result = await Customer_accountModel.get(req.params.id);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  /**
  * customer_accountModel.list
  *   @description CRUD ACTION list
  *   @returns ARRAY OF customer_account
  *
  */
  list: async (req, res) => {
    try {
      const result = await Customer_accountModel.list();
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  
  /**
  * customer_accountModel.update
  *   @description CRUD ACTION update
  *   @param ObjectId id Id customer_account
  *   @returns customer_account
  *
  */
  update: async (req, res) => {
    try {
      const result = await Customer_accountModel.update(req.body);
      res.json(result);
    } catch (err) {
      const safeErr = ErrorManager.getSafeError(err);
      res.status(safeErr.status).json(safeErr);
    }
  },
  
  
  // Custom APIs

};

export default generatedControllers;